requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.1
lazy_loader==0.3
MouseInfo==0.1.3
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.7
pytweening==1.0.7
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
ssim==0.3.0
tifffile==2023.12.9
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.4

